Tests part of the rendering pipeline.
